primetrade_backend/pyproject.toml
[project]
name = "primetrade"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "fastapi>=0.124.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "redis[async]>=7.1.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
